+++ crunch/crnlib/crn_vector.cpp
@@ -23,7 +23,7 @@ namespace crnlib
 
       size_t new_capacity = min_new_capacity;
 	  if ((grow_hint) && (!math::is_power_of_2((uint64)new_capacity)))
-		  new_capacity = math::next_pow2((uint64)new_capacity);
+		  new_capacity = (size_t)math::next_pow2((uint64)new_capacity);
 
       CRNLIB_ASSERT(new_capacity && (new_capacity > m_capacity));
